# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MoneyTamNgan API
  description: >
    JSON REST API contract between the Next.js frontend and the Node.js backend,
    covering the full system so FE/BE only need to design this once.

    Endpoints are tagged by status (see each tag's description). "Current"
    endpoints are the committed, build-now target (TOR ingestion + software/
    non-software classification, FR-1.1/FR-1.2); everything else is a
    forward-looking draft that MAY change once its requirements are worked
    out in detail — treat it as a starting point for later contract review,
    not a frozen spec.
  version: "0.2.0-draft"
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:4000/api/v1
    description: Local development (backend port TBD by BE dev)
security: []
...
paths:
...
  /tors/{id}/anomalies:
    get:
      summary: Get budget and qualification anomaly flags for a TOR
      description: Backs the clause-level highlight UI and budget warning badge.
      tags: [Anomaly Detection]
      operationId: getTorAnomalies
      parameters:
        - $ref: "#/components/parameters/TorIdParam"
      responses:
        "200":
          description: Anomaly detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TorApiAnomalyReport"
        "404":
          description: TOR not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: mtn_session
      description: >
        Planned. Session cookie set after successful login via /auth/google
        or /auth/login. Admin-only endpoints (all /admin/* paths) will also
        require this once auth lands; they're unauthenticated for now since
        no login flow exists yet.
  parameters:
    TorIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: "TOR-2569-014"
...
  schemas:
...
    TorApiAnomaly:
      oneOf:
        - type: object
          required: [type, budgetDeviationMultiplier]
          properties:
            type: { type: string, enum: [high_budget] }
            budgetDeviationMultiplier: { type: number }
        - type: object
          required: [type, clauseText, reason]
          properties:
            type: { type: string, enum: [flagged_clause] }
            clauseText: { type: string }
            reason: { type: string }
            page: { type: integer, minimum: 1 }
...
    TorApiAnomalyReport:
      type: object
      required: [id, highBudgetFlag, budgetDeviationMultiplier, anomalies]
      properties:
        id: { type: string }
        highBudgetFlag: { type: boolean }
        budgetDeviationMultiplier: { type: number }
        anomalies:
          type: array
          items:
            $ref: "#/components/schemas/TorApiAnomaly"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "TOR_NOT_FOUND"
            message:
              type: string
              example: "No TOR exists with the given id."

# Status of this file:
# - Current paths (TORs, Admin - Ingestion, Admin - Classification, /health) are
#   the committed, build-now contract.
# - All other tags/paths are a full-system draft so FE/BE only design once.
#   They are NOT locked - revisit and adjust each one once it's actually being
#   built, when real requirements/data shapes are known.